This unsigned sort is approximately twice as fast on my system according to your existing benchmarks.

Before:

```
test 0
sorted in 17.841520925 secs
compared
test 1
sorted in 19.546578784 secs
compared
test 2
sorted in 18.674980217 secs
compared
test 3
sorted in 18.138397684 secs
compared
```

After:

```
test 0
sorted in 7.32589919 secs
compared
test 1
sorted in 7.90555295 secs
compared
test 2
sorted in 7.055289701 secs
compared
test 3
sorted in 7.201343685 secs
compared
```